Changes to buttons

+++ src/edit-a-quiz/EditQuestions.jsx
@@ -6,7 +6,7 @@ const EditQuestions = ({ quiz }) => {
   const nav = useNavigate()
 
   // const questionObject = quiz.questions[index]
-  const [questionObject, setQuestionObject] = useState(quiz.questions[0])
+  const [questionObject, setQuestionObject] = useState(quiz.questions[index])
   const [question, setQuestion] = useState(questionObject.question)
   const [confirm, setConfirm] = useState(false)
   const [correctAnswer, setCorrectAnswer] = useState(questionObject.correctAnswer)
@@ -21,11 +21,11 @@ const EditQuestions = ({ quiz }) => {
 
     // creates a new question object to send back to server  
     const updatedQuestion = {
-      question, 
-      correctAnswer,
+      question: question, 
+      correctAnswer: correctAnswer,
       incorrectAnswers: [incorrectAnswer1, incorrectAnswer2, incorrectAnswer3]
     }
-    
+    console.log(updatedQuestion, 'updated')
     // fetch to API
     const res = await fetch(`https://quiz-app-server-production-09e8.up.railway.app/questions/${questionObject._id}`, {
       method: 'PUT',
@@ -35,39 +35,41 @@ const EditQuestions = ({ quiz }) => {
       },
       body: JSON.stringify(updatedQuestion)
     })
-
-    const data = await res.json()
-    console.log(data)
-
-    // need to add logic to handle errors
-    alert('Question updated successfully')
   }
 
 
   // updates index when user click next
-  function handleClickNext(event) {
+  function handleClickSaveNext(event) {
     event.preventDefault()
     setIndex(index+1)
     setQuestionObject(quiz.questions[index+1])
-  }
-
-  // handle update question
-  function handleUpdate(event) {
-    event.preventDefault()
+    setQuestion(quiz.questions[index+1].question)
+    setCorrectAnswer(quiz.questions[index+1].correctAnswer)
+    setIncorrectAnswer1(quiz.questions[index+1].incorrectAnswers[0])
+    setIncorrectAnswer2(quiz.questions[index+1].incorrectAnswers[1])
+    setIncorrectAnswer3(quiz.questions[index+1].incorrectAnswers[2])
     updateQuestion()
   }
 
-  // updates index when user click next
+  // handle update question
   function handleSubmit(event) {
     event.preventDefault()
-    // updateQuestion()
+    updateQuestion()
     nav('/quizzes')
+  }
+
+  // updates index when user click next
+  function handleSaveQuestion(event) {
+    event.preventDefault()
+    addQuestion()
+    setNewQuestion(false)
   }
 
   // handle click delete
   function handleClickDelete(event) {
     event.preventDefault()
     setConfirm(true)
+    setIndex(index+1)
   }
 
 
@@ -119,18 +121,25 @@ const EditQuestions = ({ quiz }) => {
   // handles confirming deletion of a question
   const handleConfirmDelete = async (event) => {
     event.preventDefault()
-    try {
-      await fetch(`https://quiz-app-server-production-09e8.up.railway.app/questions/${questionObject._id}`, {
-          method: "DELETE"
-      })
+    if (quiz.questions.length === 1) {
+      alert('Cannot delete this question. Quiz is required to have at least one question.')
+    } else {
+      try {
+        await fetch(`https://quiz-app-server-production-09e8.up.railway.app/questions/${questionObject._id}`, {
+            method: "DELETE"
+        })
+        setIndex(index+1)
+        setConfirm(false)
+      } catch (error) {
+        console.log(error)
+      }
+      // index === quiz.questions.length-1 ? resetForm() : setIndex(index+1) 
       setIndex(index+1)
-      setConfirm(false)
-    } catch (error) {
-      console.log(error)
+      setQuestionObject(quiz.questions[index+1])
+      setConfirm(false)``
     }
   }
-  console.log(questionObject)
-  
+
 
   // handles click on delete a question
   const confirmForm = () => {
@@ -138,7 +147,7 @@ const EditQuestions = ({ quiz }) => {
       <>
         <p> Do you want to delete this question? </p>
         <button onClick={handleConfirmDelete}> Confirm </button>
-        <button> Cancel </button>
+        <button onClick={ () => setConfirm(false) }> Cancel </button>
         <br />
       </>
     )
@@ -155,7 +164,7 @@ const EditQuestions = ({ quiz }) => {
       incorrectAnswers: [incorrectAnswer1, incorrectAnswer2, incorrectAnswer3]
     }
     // Post new question to API
-    await fetch('https://quiz-app-server-production-09e8.up.railway.app/questions', {
+    const res = await fetch('https://quiz-app-server-production-09e8.up.railway.app/questions', {
       method: 'POST',
       headers: {
         Accept: 'application/json',
@@ -163,20 +172,14 @@ const EditQuestions = ({ quiz }) => {
       },
       body: JSON.stringify(newQuestion)
     })
+    const data = await res.json()
     setNewQuestion(false)
+    quiz.questions.push(data)
+    alert('Question added successfully')
   } 
 
 
-  //render for last question
-  const render = () => {
-    return (
-      <>
-        <button onClick={ handleAddQuestion }> Add a new question </button> 
-        <button onClick={ handleSubmit }> Submit </button> 
-      </>
-    )
-  }
-  console.log(questionObject._id)
+console.log(quiz)
 
   return (
     <>
@@ -197,7 +200,7 @@ const EditQuestions = ({ quiz }) => {
               <input 
                 type='text'
                 defaultValue={questionObject.correctAnswer}
-                // value={questionObject.correctAnswer}
+                // value={questionObject.correctAnswer}  
                 onChange={(e) => setCorrectAnswer(e.target.value)}
               />
             </label>
@@ -224,22 +227,16 @@ const EditQuestions = ({ quiz }) => {
               // value={questionObject.incorrectAnswers[2]}
               onChange={(e) => setIncorrectAnswer3(e.target.value)} 
               />
-
-            { confirm && confirmForm() }
-            { newQuestion ? <button onClick={ handleConfirmAdd }> Add question </button> :
-              <>
-                <div>
-                  <button onClick={ handleUpdate }> Save changes </button>
-                  {/* <button onClick={ handleAddQuestion }> Add a new question </button> */}
-                  <button onClick={ handleClickDelete }> Delete this question </button>
-                </div>
-                { index < quiz.questions.length-1 ? 
-                  <button onClick={ handleClickNext }> Next </button> : 
-                  render()
-                }
-              </>
-            }     
+            
         </form>
+        { confirm && confirmForm() }
+        <button onClick={() => nav('/edit-a-quiz')}>Quit</button> 
+        { (index < quiz.questions.length-1 && !confirm) && <button onClick={ handleClickSaveNext }> Save & Next </button> }
+        { newQuestion && <button onClick={ handleConfirmAdd }> Save Question </button> } 
+        { (!confirm && !newQuestion) && <button onClick={ handleClickDelete }> Delete this question </button> }  
+        { (index === quiz.questions.length-1 && !confirm) &&  <button onClick={ handleAddQuestion }> Add a new question </button>}
+        { index === quiz.questions.length-1 && <button onClick={handleSubmit}> Submit </button> }
+
     </>
   )
 }
